requirements.txt
fastapi==0.110.1
pydantic==2.7.0
pydantic_settings==2.2.1
Requests==2.31.0
uvicorn==0.29.0
loguru==0.7.2
PyJWT==2.8.0
pyjwt[crypto]==2.8.0
python-dotenv==1.0.1